Update README: 'Advance Vector' -> 'Solar Guard'
Updated the game's title in the README from 'Advance Vector' to 'Solar
Guard'. 'Advance' referred to the GBA while 'Vector' referred to the
intention of having vectors for space ship movement (it was determined
from the start that this would be a space fighter game); that is, that
rather than simple movement in a vertical or horizontal scroller, there
would be acceleration/deceleration applied affecting the ship's current
momentum's velocity and direction in a 360° space. The new name
refrences the premise of the game now that one has been thought of.