Add 4 Music Tracks; It's a Music Player For Now
Additional music tracks to use in the game have been added.
Tracks starting with theme_ are intended for menus.
Tracks starting with solar_ are intended for gameplay.

There are now 6 scenes, one for each track, used to let the player play
each track (the game is basically a music player at this stage).