Make Going Back to Title Screen Possible
Executing the title screen scene again is now one of the switch cases in
the main game loop. Pressing B on the main menu will set the selection
variable used by the switch case to 0, indicating that pressing B to go
back was done as opposed to pressing A to execute the currently selected
option. Also the A button is now used where START was used except for
the title screen which still uses START.