Add Menu Logo Sprite & Implement Mission Selection
A small version of the game logo has been added to the main menu in
place of the "Solar Guard (Development Version)" text previously
displayed at the top of the screen while in the main menu.

The mission selection sub-menu now actually works, switching the game to
one of the five separate mission scenes. Of course, gameplay code will
need to be written in separate functions, no sane codebase would
duplicate gameplay code across each level/mission of a game.