Ship Turning with Movement in Facing Direction
Moved on from four-directional movement to a basic implementation of
steering the ship by having the left and right d-pad keys turn the ship
and the up and down d-pad keys move the ship forwards/backwards relative
to the direction the ship is facing.

Had to orient the ship sprite graphics to facing rightwards rather than
upwards to align with the sprite rotation logic which has 0 degrees
facing rightwards.