Mission Briefings and Backgrounds for M1 and M5
A mission briefing screen now appears when a mission is selected.
The game proceeds to the mission's gameplay when the player presses A to
proceed from the mission briefing.
The compact version of the HUD font (that the mission briefing screen
uses for the briefing text) has had some capital letters adjusted to be
wider than the standard 3 pixels to display the letter properly.

Added backgrounds for mission 1 and mission 5.
Mission 1 occurs within Earth's gravitational field.
Mission 5 occurs dangerously close to the Sun.