Implement Player Ship as C++ Object
The position and rotation of the player is now managed by variables in a
C++ class rather than the in-built position and rotation variables of a
Butano sprite. Of course, the sprite variables are still updated but
they are simply set to the values of the corresponding C++ class
variables. Which ship is selected is now an determined by an argument
given to the class constructor rather than just being determined by
which ship sprite was used in the mission's code. Gameplay and HUD
drawing functions have been updated to work with the player object
rather than with the sprite. The camera is still attached to the sprite,
but the sprite's pointer is accessed as a variable of the C++ class.