Proper HUD Implementation & Turning-Based Movement
Many HUD sprites have been added and are now displayed on-screen during
gameplay. So far only the speed, direction, coordinates, and map are
functional. The rest of the HUD elements will have their functionality
implemented as the respective gameplay functionality they relate to is
implemented. Each map has defined x and y axis limits which determine
where the indicator of the player's current position is placed on the
map, allowing the player to know where in the mission's allowable space
the player is located. The HUD font has been replicated to having
multiple differently coloured variations of itself so that text can be
drawn in various different colours in addition to green.

The player ship now has maintained speed that is accelerated and
decelerated by using the up and down keys on the D-pad respectively.
Rotating the ship doesn't affect the speed and the speed applies in the
direction the ship is facing. So the player effectively has "turning-
based movement" at the moment. The plan is to upgrade this to more
proper vector-based movement direction and velocity alteration later but
this sufficed for testing the speed meter on the HUD (the only meter
currently functioning).