Implement Buggy Vector Based Movement
Vector based movement has been implemented. However it is buggy since
the ship won't always reach it's maximum speed or change it's movement
direction if the ship already has movement in a particular direction and
the angle the ship is facing in (also the angle it accelerates in) is
not different enough to the existing movement angle. This may have to do
with how the ship's current momentum is stored as a magnitude (speed)
and direction, but is converted to separate x and y speeds to do the
application of movement change and then converted back to a magnitude
and direction, then again being converted to an x and y speed to apply
the momentum to the position of the player. This back and forth is
possibly causing loss of accuracy in the calculations. The player object
likely should just store x and y speed and work only with that. The only
place a magnitude and direction needs to be determined is the HUD for
display to the player.