Re-Implemented Player Momentum as X and Y Speeds
Player motion change is now no longer buggy in the sense that it will
properly add momentum irrespective of how different the facing angle of
the ship is to the current momentum angle.

The current implementation of limiting the overall speed of the ship is
a bit buggy (but it works while allowing the momentum direction to
change if the player accelerates while at max speed but also while
facing direction doesn't match current momentum direction). The
implementation of deceleration (allowing the player to arbitrarily
reduce speed without using the engine to propel in the opposing
direction of the current momentum) is also a currently bit buggy.