Improve Max Speed Limiting & Deceleration
The maximum speed enforcement is now much less buggy, only very slight
speed instability (significantly reduced from before) is present when
the player keeps accelerating while at maximum speed. This is due to
using a new method of changing the magnitude of the current vector from:
https://stackoverflow.com/questions/41317291/setting-the-magnitude-of-a-2d-vector

The deceleration now works flawlessly because of the use of the
magnitude changing method for deceleration but also because of a check
that makes sure when the player's overall speed becomes really small
(smaller than the deceleration per frame amount), the speed is set to
zero rather than a deceleration amount larger than the current value
just making the value a non-zero value on the other side zero (e.g. make
a positive negative or vice versa).

Also the current magnitude and direction is determined from the player's
stored x and y axis speeds once is the player object and stored
in player_ship variables so that the HUD can now just access the results
rather than wasting CPU recalculating the same values in the same frame.