Implement Afterburner & HUD's Engine Indicator
The afterburner (essentially a engine boost mode in the game) has been
implemented. Pressing the acceleration (engine thrust) and deceleration
(general speed reduction) buttons will engage the afterburner which, for
now, provided double the rate of acceleration that the normal engine use
provides. A basic ship animation in which one of the two sprites showing
the ship using afterburners is switched to during afterburner is used.
The original intent was for the ship to rapidly switch between both
of a given ship's two afterburner use sprites to animate the afterburner
appearance but my initial attempts at using Butano's animation actions,
specifically create_sprite_animate_action_forever() resulted in a
duplicate of the ship sprite that didn't rotate appearing alongside the
existing ship. Solving this is on the backburner since the simpler
animation it has now is suffice for now. Fixing this is a nice-to-have
but not a must-have.

The HUD's engine status indicator now works, showing whether the player
is accelerating, decelerating, using afterburner, or doing none of the
aforementioned (idle).