Add Folders and READMEs for OGA Artwork and OST
Copies of art from Open Game Art and relevant licensing information have
been added to the 'Open Game Art' folder. This folder also contains
copies of game art that has been derived from art obtained from Open
Game Art and licenses these derivatives using the same license as the
original art (for some licenses, this is required, but this is done for
all derivative art regardless of whether the original art's license
required it or not).

Exports of Beepbox song data in JSON format for the music tracks have
been added to the 'Original Soundtrack' folder, along with a README
containing information about the music.

The actual Butano project files have been moved from the root of the
repository inside the 'Solar_Guard' folder. This is so that the new
folders (and any other future content that isn't a part of the Butano
project) can be added to the repository without awkwardly sitting around
in a sea of Butano project related folders and files.