Engine Fuel & Overheat + Further HUD Implementation
The engine fuel consumption and heating mechanics have been implemented.
When the engine is used, fuel is consumed and heat is generated.
Using afterburner uses fuel quicker and heats the engine quicker.
If fuel runs out the engine can no longer be used.
If the engine heat reaches 100%, it cannot be used until it has cooled
back down to 0%, at which point it can be used again.

The HUD has been further worked on, it now actually displays the shield,
health, fuel, and engine heat levels from the player ship's variables.
The target shields and health section now says NO TGT (no target) rather
than a fake 100% reading for both the target shields and health. This
part of the HUD will be made functional once targeting is implemented.
Likewise the meters on the right side of the screen for utility energy
and utility heat will be implemented when ship weapons/utilities are
implemented.