Add Space Junk and Asteroids to Missions 1 and 2
Space junk is now present in mission 1.
Asteroids are present in mission 2.
The space junk and asteroids exists as instances of a C++ class.
These instances have their pointers in a bn::array so that they can be
accessed systematically for any possible future targeting functionality
(the player ship being able to "switch targets" to see the status of
different entities on the HUD).

Also fixed spr_space_junk.json's sprite height value from 8 to 16.