Restore Player Animation and Add Lasers to Missions 2-5
The player animation code was moved out of the player's movement
function into it's own function. This resulted in the player sprite no
longer animating in missions 2-5 since the player's animation function
was not added to these missions' code loops.

The player's lasers were not in missions 2-5 since the new function for
firing them wasn't added to the relevant code loops, nor were the laser
objects created in the functions for these missions.

The relevant animation and laser code has been added to the code loops
of missions 2-5 so that the player animates again and now can fire
lasers in missions 2-5.