Implement Radar and Mission 2 Moving Transport
The radar now works, showing the position of nearby space junk,
asteroids, and allies. The allies also show up on the map as well.

The transport is now present in mission 2 and moves from the bottom-left
corner of the mission area to the top-right corner of the mission area.
For now there's no mission complete trigger, the transport just moves
to it's intended destination in the map area and then stops.