Change HUD Objective to Sprites + Add M2 Objective to HUD
The objective is now displayed using sprites in the graphics folder
rather than using the text sprite generation functionality of Butano.
This is for the same reason the most of the rest of the HUD is also
drawn this way: it uses much less CPU to draw it this way.

The mission 2's objective is now displayed on the HUD during mission 2.