Update Ship Select to Show Burn Rate in L/s
The ships use fuel at a rate of 0.01/frame going forward and half that
at 0.005/frame going in reverse. Multiplying this by 60 to go from per
frame to per second and multiplying that again by 100 to do the simple
"litres version" of how much fuel is present results in a forward rate
of 60Litres/second and a reverse rate of 30Litres/second. Whether this
is in any way a realistic rate fuel consumption for afterburning engines
if irrelevant to Solar Guard since the fuel consumption rate relative to
the amount of fuel available is balanced based on gameplay rather than
realism since Solar Guard is a game, not a simulator. In fact engines
that need air wouldn't work in space so the engines on these space ships
wouldn't just simply be the same as regular modern jet engines anyway.

Also made the Space Fighter Mark III and the Stable Space Fighter use
60L/s for reverse engines since those are about as large (on the sprite
animation) and just as powerful (in gameplay) as the standard engines on
those craft (most craft have reverse engines that are smaller and less
powerful than their standard engines).