Implement Mission 5's Civilian Ship and Escape Pods
The civilian ship that has systems failure near the sun is now present.
There are three escape pods that have jettisoned from the ship. The ship
and the escape pods are drifting towards the sun. The player can pick up
the escape pods and drop them off to the carrier by colliding with an
escape pod and then colliding with the carrier when carrying the escape
pod. The objective is to successfully retrieve the escape pods before
they fly outside of the mission area.