Shield/Hull Regen + Mission 2 Asteroid Collision
The player's shield and hull now recover (as per their stats shown on
the ship select screen, not all ships have hull recovery).

The player can now collide with asteroids in mission 2.
Doing so will cause 40% damage to the player and destroy the asteroid.

Also the asteroids now appear on the map as well as the radar.