Space Junk Collision Now Causes Damage
Colliding into the space junk will now cause 15% damage to the player
and destroy the space junk.

It is possible to fail the mission if all hull integrity is lost, but
due to the low damage per collision and regenerating shields, you'd have
to intentionally put effort into colliding with enough space junk fast
enough to do that.

Also, each piece of space junk appears on the map now.