Changed Asteroid Collision Function to Use bn::abs()
Subtracting two x position values and checking how large the absolute
value is was implemented as an alternative method of collision checking
to see if it stopped the slow down of the game that occurred when firing
lasers. It didn't work, since the game still slows down when lasers are
fired in mission 2.