Adjust Pirate Space Fighter Ship Select Screen
Added the acceleration stats that were missing.

This screen is for mission 4, in which the player uses a captured pirate
space fighter to perform recon on pirate bases without being detected.

This mission won't be functional by the time GBA JAM 2021 ends, so the
version of the game submitted to the jam won't have this screen viewable
in-game since mission 4 still has the standard ship select code.