Add Pause Menu
Implemented a simple pause menu that lets the player resume or exit to
the main menu.

This is activated by pressing START and replaces the instant exit that
was activated by pressing SELECT (pressing SELECT now does nothing).

Actually thought of a proper way of stopping the player from firing a
laser from a menu selection instance of pressing A: Wait until A is no
longer held before activating the menu selection when A is pressed
(involves using the bn::keypad::a_held() function rather than using the
bn::keypad::a_pressed() function).