Remove HUD Target Display & Move HUD Speed Rightwards
The targeting functionality (which would show the shields/hull of a
targeted entity on the HUD) hasn't been implemented in time for the GBA
JAM 2021 submission. Therefore, the targeting display has been removed
from the HUD and the speed/direction display moved rightwards from the
top-centre to the top-right where the target information was.