Update Mission Briefing Text, Including Notes/Issues
The mission briefing text for each mission has been revised.

The fact that slowdown occurs in mission 2 when firing lasers has been
added to mission 2's briefing to inform the player of this.

The fact that missions 3 & 4 are not implemented has been noted in their
respective mission briefings to inform the player of this.