Fix Ships Hull-Repairing Fast When They're Not Meant To
Ships which weren't meant to have hull self-repair capability actually
had it and it was fast. The reason is that setting the
hull_recharge_portion to 0 indicates that the ship does not have this
function. The hull_recharge_rate which indicates how many frames of
delay is in-between each hull point recovered was set to 0 for ships
that were not meant to have hull self-repair but their
hull_recharge_portion was still set to 20 (meaning 20%, indicating that
hull repairs up to the nearest fifth). This resulted in the ships still
repairing within the current fifth of hull integrity really fast at a
rate of 1% hull a frame (which is 60% hull a second since the game runs
at 60FPS) since the variable specifying delay was set to 0 under the
assumption that the recharge portion variable would also be set to 0
(therefore making the delay value inconsequential since hull-repair
doesn't occur at all) but it was not. This has been fixed.