Adjust Wording About Mission 2 Slowdown Issue
It appears that the slowdown when lasers are fired in mission 2 was made
somewhat rare by the asteroid distance check that occurs to determine if
collision checking should occur. In my testing, you can fire lasers and
slowdown won't occur a lot of the time, just occasionally.

The wording of the relevant text in mission 2's briefing and the known
issues page has been adjusted to speak of this slow down as possible
rather than certain due to this.