Remove Player Coordinates from HUD
The coordinates display was there for development purposes.
They aren't a part of the HUD the player is meant to use.
Therefore, they've been removed by having the code commented out for the
jam submission version of the game.
They can be re-added afterwards at any stage if they're needed for
development purposes again.