Transport Accelerates from a Speed of Zero
When mission 2 begins, the transport now starts with a speed of zero and
takes a few seconds to accelerate to the speed it will have throughout
the level.

Not only does the acceleration give the transport a nice take-off
sequence, but it also gives the player a few more seconds to react to
seeing the transport taking off and catch up.