Make Shield/Hull Number Grey When 00%
Normally the number is the colour of the current colour region the bar
is filled up to. Previosly the number was just the colour of the lowest
shield/health bar colour region when at 00%. Now the number turns grey
when at 00% to make further visual distinction between low shields/hull
and no shields/hull.