Make Shield/Hull Bar Notch Present with Part of 4%
Each notch on the shield and hull bars is equal to 4% (there are 25
notches). Previously a notch would only be present in if the whole 4% of
the shields/hull it represented was present (e.g. present when
shields/hull at 80%, depleted when at 79%). Now the notch will be
present if any part of the 4% is present (e.g. not only present when
shields/hull at 80%, but also present when at 77%, 78%, and 79%, only
depleting once shields/hull falls to 76% or below).